Horizontal Chart 작업중

+++ app/src/main/java/kr/co/carboncheck/android/carboncheckapp/dataobject/MemberUsageData.kt
@@ -3,6 +3,6 @@ package kr.co.carboncheck.android.carboncheckapp.dataobject
 // total_usage_fragment Recycler view 의 아이템
 data class MemberUsageData (
     val user_name: String,
-    val target_usage: Int,
-    val current_usage: Int,
+    val target_usage: Float,
+    val current_usage: Float
 )
